% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{normalize_sim2_profile}
\alias{normalize_sim2_profile}
\title{Normalize and validate a Simulation 2 profile name.}
\usage{
normalize_sim2_profile(profile)
}
\arguments{
\item{profile}{Character profile name.}
}
\value{
One lower-case profile name.
}
\description{
Normalize and validate a Simulation 2 profile name.
}
