% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_scheduler.R
\name{shutdown_global_hal_worker}
\alias{shutdown_global_hal_worker}
\title{Release process-local resources owned by one persistent HAL slot.}
\usage{
shutdown_global_hal_worker()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Release process-local resources owned by one persistent HAL slot.
}
