% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.popart_fit}
\alias{summary.popart_fit}
\title{Summarize a PopART fit}
\usage{
\method{summary}{popart_fit}(object, ...)
}
\arguments{
\item{object}{A fitted \code{popart_fit} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
An object of class \code{summary.popart_fit}.
}
\description{
Summarize a PopART fit
}
