% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw_fit.R
\name{fit_ipw}
\alias{fit_ipw}
\title{Fit a parametric IPW estimator}
\usage{
fit_ipw(data, weight_formula, version = c("naive", "proposed"))
}
\arguments{
\item{data}{A combined trial and auxiliary data frame containing:
\itemize{
\item{\code{S}: trial-membership indicator;}
\item{\code{R}: response indicator;}
\item{\code{C}: censoring indicator;}
\item{\code{A}: treatment indicator;}
\item{\code{Y}: outcome;}
\item{\code{wt}: analysis weight;}
\item{the variables in \code{weight_formula}.}
}}

\item{weight_formula}{A censoring formula for the naive estimator or combined
sample-membership formula for the proposed estimator.}

\item{version}{A character string equal to \code{"naive"} or
\code{"proposed"}.}
}
\value{
A one-row data frame containing:
\itemize{
\item{\code{etahat_0}: estimated mean potential outcome under control;}
\item{\code{etahat_1}: estimated mean potential outcome under treatment;}
\item{\code{cov_00} and \code{cov_11}: estimated variances;}
\item{\code{cov_01}: estimated covariance.}
}
}
\description{
Fits either the naive or proposed parametric inverse-probability weighted
estimator for the two mean potential outcomes.
}
\keyword{internal}
