% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popart.R
\docType{package}
\name{popart-package}
\alias{popart}
\alias{popart-package}
\title{popart: Population-Augmented Analysis of Randomized Trials}
\description{
The package fits naive and proposed g-formula, IPW, and AIPW estimators.
}
\author{
\strong{Maintainer}: PopART Team \email{popart.package@example.com}

}
\keyword{internal}
